Refactor step 2, ready for multiple streamline generators
