+++ ApplicationLibCode/ProjectDataModel/Streamlines/CMakeLists_files.cmake
@@ -3,6 +3,7 @@ set (SOURCE_GROUP_HEADER_FILES
 ${CMAKE_CURRENT_LIST_DIR}/RimStreamline.h
 ${CMAKE_CURRENT_LIST_DIR}/RimStreamlineInViewCollection.h
 ${CMAKE_CURRENT_LIST_DIR}/StreamlineGenerator.h
+${CMAKE_CURRENT_LIST_DIR}/StreamlineGeneratorBase.h
 ${CMAKE_CURRENT_LIST_DIR}/StreamlineDataAccess.h
 )
 
@@ -10,6 +11,7 @@ set (SOURCE_GROUP_SOURCE_FILES
 ${CMAKE_CURRENT_LIST_DIR}/RimStreamline.cpp
 ${CMAKE_CURRENT_LIST_DIR}/RimStreamlineInViewCollection.cpp
 ${CMAKE_CURRENT_LIST_DIR}/StreamlineGenerator.cpp
+${CMAKE_CURRENT_LIST_DIR}/StreamlineGeneratorBase.cpp
 ${CMAKE_CURRENT_LIST_DIR}/StreamlineDataAccess.cpp
 )
